RESOLVED: #0029520: The whole screen turns yellow and the saving progress dialog keeps flashing while muvee was saving to 1080p using cubetwist
Ok it looks like we have a winner. The problem was that even though we were using an FBO for saving, we were still initializing the Hidden window that was not being used. This was a rather substantial resource allocation for an unused code path. And that would cause mid-range machines that have support for FBOs to run out of VRAM. What the previous commit has done is to not allocate any resource to the hidden window if we are using FBOs. Et voila!

This commit does some minor code reformatting and sets the multisampling rate of FBOs to 8 if the opengl version is 2.1 and above.