#0029582: Saving muvee to Computer 720p or 1080p format would results the muvee to be distorted
We specifically set texture sizes to 1k if the graphics card is "ATI Radeon HD 2400 XT OpenGL Engine".

This is because we know that is has 128MB of vram but when we programatically check the vram. 256MB is returned. So I've been left with no choice but to specifically seek this string. In the future, this check might go away if the drivers are fixed.