#0030047: [OEM - CoGen] New Animations for Credits
I will now revert most of the things I removed in TextEffect_RollingSummary. I had revoved it because there seemed to be a mismatch in the enums vs the list in the combo box. I have narrowed it down to in the enum inside MVPersonalize.xaml.cs (which btw I am not reverting). All the other files, however, can be reverted to their orignal content. The reason we want to minimize the change is to promote backward compatibility.