0030050: [OEM - CoGen] A style dedicated to create rolling credits
Removed alpha portion from the color style param, thereby implicitly assume alpha as 0xff. For some reasons, when we specify color with alpha values in the style parameters, the colors will be incorrectly converted (Check /muveeDebug/AppInfoOut.txt.) This bug will be addressed in the future. In the meantime, style authors should continue to avoid specifying alpha in color style params.

Changed English description from Normal/Tall to Narrow/Wide. This is a more accurate translation from the Japanese string. It is also acceptable to describe line spacing as narrow and wide.