0032345: New construction mode for Cinnamon
(The actual name for this mode is yet to be finalized. We shall call it "summary mode" or "summarize shots" tentatively.)

Created a separated code path for the new construction mode. There is no UI to switch to this mode yet. It is currently accessible via C:\muveeDebug\MakeSummarizeShots.txt.

Made my first crack at the new construction mode, based on KIA mode but it must fit to music:
- With too much visuals, it drops the last media to fit the required music duration.
- With too few visuals, it stretches the last media to fit the required music duration.

Added code to check for last anchor segment, and propagated this down and made this available at the deeper levels of construction logic.