0032345: New construction mode for Cinnamon
Implemented an algorithm to find an optimal segment trimming and warping amount such that strong beats are in the middle of all transitions.

Trimming and warping limits can be set at the stylelib level. By default, the warping limits are 95% and 105%, and the maximum trimming is at 25%.

Moved TVPairs typedef to VirtualMusicDescriptor. Cleaned up DescriptorLingo and VirtualMusicDescriptors.